- Consolidate LocalDeepCopy unit tests - Fix bug of using offset for block transfers in MPI
